feature36: Product Server 숙박업체 재고 조회 Service - booking -> product call 테스트코드 확인 - 샘플 데이터 (Test h2 DB) 용 .sql 파일 추가 - Product com->org package rename
