+++ expedia-server/booking-server/build.gradle
@@ -1,0 +1,7 @@
+jar {
+    enabled = false
+}
+
+bootJar {
+    enabled = true
+}
