+++ expedia-server/booking-server/src/main/java/org/clonecoder/bookingserver/domain/booking/BookingService.java
@@ -1,6 +1,5 @@
 package org.clonecoder.bookingserver.domain.booking;
 
-import com.clonecoder.productserver.domain.accommodation.AccommodationService;
 import lombok.RequiredArgsConstructor;
 import org.clonecoder.bookingserver.domain.Booking;
 import org.clonecoder.bookingserver.domain.BookingGuests;
