+++ expedia-server/booking-server/src/main/java/org/clonecoder/bookingserver/infrastructure/BookingStoreImpl.java
@@ -1,6 +1,6 @@
 package org.clonecoder.bookingserver.infrastructure;
 
-import com.clonecoder.productserver.domain.accommodation.AccommodationService;
+import org.clonecoder.productserver.domain.accommodation.AccommodationService;
 import lombok.RequiredArgsConstructor;
 import org.clonecoder.bookingserver.domain.Booking;
 import org.clonecoder.bookingserver.domain.BookingGuests;
