--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/ProductServerApplication.java
@@ -1,13 +1,0 @@
-package com.clonecoder.productserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProductServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProductServerApplication.class, args);
-    }
-
-}
