--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/domain/AccommodationRole.java
@@ -1,31 +1,0 @@
-package com.clonecoder.productserver.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class AccommodationRole {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "accommodation_id")
-    private Accommodation accommodation;
-
-    private String progStatusCd;
-    private String stoppedDts;
-    private String stoppedRsn;
-
-    private String lastModifiedBy;
-    private LocalDateTime lastModifiedAt;
-    private String createdBy;
-    private LocalDateTime createdAt;
-}
