--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/domain/AccommodationRoomFee.java
@@ -1,32 +1,0 @@
-package com.clonecoder.productserver.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-@Getter
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class AccommodationRoomFee {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "accommodation_room_id")
-    private AccommodationRoom accommodationRoom;
-
-    private int active;
-    private String feeName;
-    private BigDecimal fee;
-
-    private String lastModifiedBy;
-    private LocalDateTime lastModifiedAt;
-    private String createdBy;
-    private LocalDateTime createdAt;
-}
