--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/domain/accommodation/AccommodationService.java
@@ -1,18 +1,0 @@
-package com.clonecoder.productserver.domain.accommodation;
-
-import com.clonecoder.productserver.domain.AccommodationRoom;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class AccommodationService {
-    private final AccommodationStore accommodationStore;
-
-    public void stock(Long accommodationRoomId) {
-        AccommodationRoom accommodationRoom = accommodationStore.getAccommodationRoom(accommodationRoomId);
-        log.info("stock() : " + accommodationRoom.getStock());
-    }
-}
