--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/domain/accommodation/AccommodationStore.java
@@ -1,7 +1,0 @@
-package com.clonecoder.productserver.domain.accommodation;
-
-import com.clonecoder.productserver.domain.AccommodationRoom;
-
-public interface AccommodationStore {
-    AccommodationRoom getAccommodationRoom(Long accommodationRoomId);
-}
