--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationCommonCodeRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.AccommodationCommonCode;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationCommonCodeRepository extends JpaRepository<AccommodationCommonCode, Long> {
-
-}
