--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationGroupCommonCodeRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.AccommodationGroupCommonCode;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationGroupCommonCodeRepository extends JpaRepository<AccommodationGroupCommonCode, Long> {
-
-}
