--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.Accommodation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationRepository extends JpaRepository<Accommodation, Long> {
-
-}
