--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationRoleRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.AccommodationRole;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationRoleRepository extends JpaRepository<AccommodationRole, Long> {
-
-}
