--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationRoomFeeRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.AccommodationRoomFee;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationRoomFeeRepository extends JpaRepository<AccommodationRoomFee, Long> {
-
-}
