--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationRoomInfoRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.AccommodationRoomInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationRoomInfoRepository extends JpaRepository<AccommodationRoomInfo, Long> {
-
-}
