--- expedia-server/product-server/src/main/java/com/clonecoder/productserver/infrastructure/AccommodationRoomRepository.java
@@ -1,8 +1,0 @@
-package com.clonecoder.productserver.infrastructure;
-
-import com.clonecoder.productserver.domain.AccommodationRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccommodationRoomRepository extends JpaRepository<AccommodationRoom, Long> {
-
-}
