+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/domain/AccommodationGroupCommonCode.java
@@ -1,0 +1,30 @@
+package org.clonecoder.productserver.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import java.time.LocalDateTime;
+
+@Getter
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+public class AccommodationGroupCommonCode {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String grpCode;
+    private String grpCodeName;
+    private String grpCodeDesc;
+
+    private String lastModifiedBy;
+    private LocalDateTime lastModifiedAt;
+    private String createdBy;
+    private LocalDateTime createdAt;
+}
