+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/domain/AccommodationRoom.java
@@ -1,0 +1,32 @@
+package org.clonecoder.productserver.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Getter
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+public class AccommodationRoom {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "accommodation_id")
+    private Accommodation accommodation;
+
+    private String roomType;
+    private String roomName;
+    private int maxEntranceCnt;
+    private int stock;
+
+    private String lastModifiedBy;
+    private LocalDateTime lastModifiedAt;
+    private String createdBy;
+    private LocalDateTime createdAt;
+}
