+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationCommonCodeRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.AccommodationCommonCode;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationCommonCodeRepository extends JpaRepository<AccommodationCommonCode, Long> {
+
+}
