+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationGroupCommonCodeRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.AccommodationGroupCommonCode;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationGroupCommonCodeRepository extends JpaRepository<AccommodationGroupCommonCode, Long> {
+
+}
