+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.Accommodation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationRepository extends JpaRepository<Accommodation, Long> {
+
+}
