+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationRoleRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.AccommodationRole;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationRoleRepository extends JpaRepository<AccommodationRole, Long> {
+
+}
