+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationRoomFeeRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.AccommodationRoomFee;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationRoomFeeRepository extends JpaRepository<AccommodationRoomFee, Long> {
+
+}
