+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationRoomInfoRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.AccommodationRoomInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationRoomInfoRepository extends JpaRepository<AccommodationRoomInfo, Long> {
+
+}
