+++ expedia-server/product-server/src/main/java/org/clonecoder/productserver/infrastructure/AccommodationRoomRepository.java
@@ -1,0 +1,8 @@
+package org.clonecoder.productserver.infrastructure;
+
+import org.clonecoder.productserver.domain.AccommodationRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccommodationRoomRepository extends JpaRepository<AccommodationRoom, Long> {
+
+}
