--- expedia-server/product-server/src/test/java/com/clonecoder/productserver/ProductServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.clonecoder.productserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProductServerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
