+++ expedia-server/product-server/src/test/java/org/clonecoder/productserver/ProductServerApplicationTests.java
@@ -1,0 +1,13 @@
+package org.clonecoder.productserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductServerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
