+++ expedia-server/product-server/src/test/resources/application.yml
@@ -1,0 +1,7 @@
+spring:
+  # H2Database
+  datasource:
+    url: jdbc:h2:mem:test
+    username: sa
+    password:
+    driver-class-name: org.h2.Driver
